Split Data input screens up to reduce clutter
