+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Setup.SetupAddCareGiverComplete"></activity>
+        <activity android:name=".Setup.SetupUserInfo4"></activity>
+        <activity android:name=".Setup.SetupUserInfo3" />
+        <activity android:name=".Setup.SetupAddCareGiverComplete" />
         <activity android:name=".Setup.SetupVerification" />
         <activity android:name=".Setup.SetupComplete" />
         <activity android:name=".Setup.SetupHighBloodSugar" />
