+++ app/src/main/res/layout/activity_setup_add_care_giver_complete.xml
@@ -18,18 +18,6 @@
         app:srcCompat="@android:color/background_light" />
 
     <ImageButton
-        android:id="@+id/backButton"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:background="@drawable/backbutton"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.17"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/backgroundImage"
-        app:layout_constraintVertical_bias="1.0" />
-
-    <ImageButton
         android:id="@+id/nextButton"
         android:layout_width="100dp"
         android:layout_height="100dp"
